pyproject.toml
[project]
name = "parsershopbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "playwright>=1.51.0",
    "psycopg>=3.2.6",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
]
